Reactor削除 & .gitgnore適用に伴うキャッシュ削除
