+++ NeutralInModes.csproj
@@ -5,27 +5,24 @@
         <DebugType>embedded</DebugType>
 
         <VersionPrefix>1.0.0</VersionPrefix>
-        <VersionSuffix>dev</VersionSuffix> 
-        <Description>Mod generated using Reactor Template</Description>
-        <!-- <Authors>oshurecat(Shure)</Authors> -->
-    </PropertyGroup>
+        <Description>Neutral In Modes</Description>
+		<Authors>oshurecat</Authors>
 
-    <PropertyGroup>
-        <GamePlatform Condition="'$(GamePlatform)' == ''">Steam</GamePlatform>
-        <GameVersion Condition="'$(GamePlatform)' == 'Steam'">2022.6.21</GameVersion>
-        <GameVersion Condition="'$(GamePlatform)' == 'Itch'">2022.6.21</GameVersion>
+		<DefineConstants>$(DefineConstants);STEAM</DefineConstants>
+		<AllowUnsafeBlocks>True</AllowUnsafeBlocks>
+		<ApplicationIcon />
+		<Win32Resource />
+		<AssemblyName>NeutralInModes</AssemblyName>
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="Reactor" Version="1.2.5" />
-        <PackageReference Include="BepInEx.IL2CPP" Version="6.0.0-be.570" />
-        <PackageReference Include="AmongUs.GameLibs.$(GamePlatform)" Version="$(GameVersion)" PrivateAssets="all" />
-
-        <PackageReference Include="BepInEx.AutoPlugin" Version="1.0.1" PrivateAssets="all" />
+        <PackageReference Include="BepInEx.IL2CPP" Version="6.0.0-be.562" />
+        <PackageReference Include="AmongUs.GameLibs.Steam" Version="2022.8.23" PrivateAssets="all" />
         <PackageReference Include="BepInEx.IL2CPP.MSBuild" Version="1.1.1" PrivateAssets="all" />
     </ItemGroup>
 
-    <Target Name="Copy" AfterTargets="Build" Condition="'$(AmongUs)' != ''">
-        <Copy SourceFiles="$(TargetPath)" DestinationFolder="$(AmongUs)/BepInEx/plugins/" UseSymboliclinksIfPossible="true" />
-    </Target>
+    <Target Name="CopyCustomContent" AfterTargets="AfterBuild" Condition="Exists('$(AmongUsLatest)/BepInEx/plugins/')">
+		<Message Text="Second occurrence" />
+		<Copy SourceFiles="$(ProjectDir)\bin\$(Configuration)\netstandard2.1\$(AssemblyName).dll" DestinationFolder="$(AmongUsLatest)/BepInEx/plugins/" />
+	</Target>
 </Project>
